chore: minor flows table refactoring
