+++ packages/react-ui/src/app/routes/flows/index.tsx
@@ -4,27 +4,18 @@ import { t } from 'i18next';
 import {
   CheckIcon,
   ChevronDown,
-  CornerUpLeft,
-  Download,
   EllipsisVertical,
   Import,
   Plus,
-  Trash2,
-  UploadCloud,
   Workflow,
 } from 'lucide-react';
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import { ConfirmationDeleteDialog } from '@/components/delete-dialog';
 import { useEmbedding, useNewWindow } from '@/components/embed-provider';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
-import {
-  BulkAction,
-  DataTable,
-  RowDataWithActions,
-} from '@/components/ui/data-table';
+import { DataTable, RowDataWithActions } from '@/components/ui/data-table';
 import { DataTableColumnHeader } from '@/components/ui/data-table/data-table-column-header';
 import {
   DropdownMenu,
@@ -33,28 +24,22 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 import { PermissionNeededTooltip } from '@/components/ui/permission-needed-tooltip';
-import { LoadingSpinner } from '@/components/ui/spinner';
 import { INTERNAL_ERROR_TOAST, toast } from '@/components/ui/use-toast';
 import { FlowStatusToggle } from '@/features/flows/components/flow-status-toggle';
 import { ImportFlowDialog } from '@/features/flows/components/import-flow-dialog';
-import { MoveFlowDialog } from '@/features/flows/components/move-flow-dialog';
 import { SelectFlowTemplateDialog } from '@/features/flows/components/select-flow-template-dialog';
+import { useFlowsBulkActions } from '@/features/flows/lib/use-flows-bulk-actions';
 import { flowsApi } from '@/features/flows/lib/flows-api';
-import { flowsUtils } from '@/features/flows/lib/flows-utils';
 import { FolderBadge } from '@/features/folders/component/folder-badge';
 import {
   FolderFilterList,
   folderIdParamName,
 } from '@/features/folders/component/folder-filter-list';
 import { foldersApi } from '@/features/folders/lib/folders-api';
-import { PushToGitDialog } from '@/features/git-sync/components/push-to-git-dialog';
-import { gitSyncHooks } from '@/features/git-sync/lib/git-sync-hooks';
 import { PieceIconList } from '@/features/pieces/components/piece-icon-list';
 import { useAuthorization } from '@/hooks/authorization-hooks';
-import { platformHooks } from '@/hooks/platform-hooks';
 import { authenticationSession } from '@/lib/authentication-session';
 import { formatUtils, NEW_FLOW_QUERY_PARAM } from '@/lib/utils';
-import { GitBranchType } from '@activepieces/ee-shared';
 import { FlowStatus, Permission, PopulatedFlow } from '@activepieces/shared';
 
 import FlowActionMenu from '../../../app/components/flow-actions-menu';
@@ -93,50 +78,6 @@ const FlowsPage = () => {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const openNewWindow = useNewWindow();
   const [searchParams] = useSearchParams();
-
-  const { platform } = platformHooks.useCurrentPlatform();
-  const { gitSync } = gitSyncHooks.useGitSync(
-    authenticationSession.getProjectId()!,
-    platform.environmentsEnabled,
-  );
-  const userHasPermissionToWriteFolder = checkAccess(Permission.WRITE_FOLDER);
-  const userHasPermissionToUpdateFlow = checkAccess(Permission.WRITE_FLOW);
-  const userHasPermissionToPushToGit = checkAccess(
-    Permission.WRITE_PROJECT_RELEASE,
-  );
-  const isDevelopmentBranch =
-    gitSync && gitSync.branchType === GitBranchType.DEVELOPMENT;
-
-  const { mutate: exportFlows, isPending: isExportPending } = useMutation({
-    mutationFn: async (flows: PopulatedFlow[]) => {
-      if (flows.length === 0) {
-        return;
-      }
-      if (flows.length === 1) {
-        await flowsUtils.downloadFlow(flows[0].id);
-        return;
-      }
-      const zip = await flowsUtils.downloadFlowsIntoZip(flows);
-      const content = await zip.generateAsync({ type: 'blob' });
-      const url = URL.createObjectURL(content);
-      const a = document.createElement('a');
-      a.style.display = 'none';
-      a.href = url;
-      a.download = 'flows.zip';
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-    },
-    onSuccess: () => {
-      toast({
-        title: t('Success'),
-        description: t('Flows have been exported.'),
-        duration: 3000,
-      });
-    },
-    onError: () => toast(INTERNAL_ERROR_TOAST),
-  });
   const projectId = authenticationSession.getProjectId()!;
   const { data, isLoading, refetch } = useQuery({
     queryKey: ['flow-table', searchParams.toString(), projectId],
@@ -373,167 +314,15 @@ const FlowsPage = () => {
     },
   ];
 
-  const bulkActions: BulkAction<PopulatedFlow>[] = useMemo(
-    () => [
-      {
-        render: (_, resetSelection) => {
-          return (
-            <div onClick={(e) => e.stopPropagation()}>
-              <DropdownMenu
-                modal={true}
-                open={isDropdownOpen}
-                onOpenChange={setIsDropdownOpen}
-              >
-                {selectedRows.length > 0 ? (
-                  <DropdownMenuTrigger asChild>
-                    <Button
-                      className="h-9 w-full"
-                      variant={'default'}
-                      onClick={() => {
-                        setIsDropdownOpen(!isDropdownOpen);
-                      }}
-                    >
-                      {selectedRows.length > 0
-                        ? `${t('Actions')} (${selectedRows.length})`
-                        : t('Actions')}
-                      <ChevronDown className="h-3 w-4 ml-2" />
-                    </Button>
-                  </DropdownMenuTrigger>
-                ) : null}
-
-                <DropdownMenuContent>
-                  <PermissionNeededTooltip
-                    hasPermission={userHasPermissionToPushToGit}
-                  >
-                    <PushToGitDialog
-                      flowIds={selectedRows.map((flow) => flow.id)}
-                    >
-                      <DropdownMenuItem
-                        disabled={!userHasPermissionToPushToGit}
-                        onSelect={(e) => {
-                          e.preventDefault();
-                        }}
-                      >
-                        <div className="flex cursor-pointer  flex-row gap-2 items-center">
-                          <UploadCloud className="h-4 w-4" />
-                          <span>{t('Push to Git')}</span>
-                        </div>
-                      </DropdownMenuItem>
-                    </PushToGitDialog>
-                  </PermissionNeededTooltip>
-                  {!embedState.hideFolders && (
-                    <PermissionNeededTooltip
-                      hasPermission={
-                        userHasPermissionToUpdateFlow ||
-                        userHasPermissionToWriteFolder
-                      }
-                    >
-                      <MoveFlowDialog
-                        flows={selectedRows}
-                        onMoveTo={() => {
-                          setRefresh(refresh + 1);
-                          resetSelection();
-                          setSelectedRows([]);
-                          refetch();
-                          setIsDropdownOpen(false);
-                        }}
-                      >
-                        <DropdownMenuItem
-                          disabled={
-                            !userHasPermissionToUpdateFlow ||
-                            !userHasPermissionToWriteFolder
-                          }
-                          onSelect={(e) => e.preventDefault()}
-                        >
-                          <div className="flex cursor-pointer  flex-row gap-2 items-center">
-                            <CornerUpLeft className="h-4 w-4" />
-                            <span>{t('Move To')}</span>
-                          </div>
-                        </DropdownMenuItem>
-                      </MoveFlowDialog>
-                    </PermissionNeededTooltip>
-                  )}
-                  <DropdownMenuItem
-                    onClick={() => {
-                      exportFlows(selectedRows);
-                      resetSelection();
-                      setSelectedRows([]);
-                      setIsDropdownOpen(false);
-                    }}
-                  >
-                    <div className="flex cursor-pointer flex-row gap-2 items-center">
-                      {isExportPending ? (
-                        <LoadingSpinner />
-                      ) : (
-                        <Download className="h-4 w-4" />
-                      )}
-                      <span>
-                        {isExportPending ? t('Exporting') : t('Export')}
-                      </span>
-                    </div>
-                  </DropdownMenuItem>
-                  <PermissionNeededTooltip
-                    hasPermission={userHasPermissionToUpdateFlow}
-                  >
-                    <ConfirmationDeleteDialog
-                      title={`${t('Delete')} Selected Flows`}
-                      message={
-                        <>
-                          <div>
-                            {t(
-                              'Are you sure you want to delete these flows? This will permanently delete the flows, all their data and any background runs.',
-                            )}
-                          </div>
-                          {isDevelopmentBranch && (
-                            <div className="font-bold mt-2">
-                              {t(
-                                'You are on a development branch, this will not delete the flows from the remote repository.',
-                              )}
-                            </div>
-                          )}
-                        </>
-                      }
-                      mutationFn={async () => {
-                        await Promise.all(
-                          selectedRows.map((flow) => flowsApi.delete(flow.id)),
-                        );
-                        setRefresh(refresh + 1);
-                        resetSelection();
-                        setSelectedRows([]);
-                        refetch();
-                        setIsDropdownOpen(false);
-                      }}
-                      entityName={t('flow')}
-                    >
-                      <DropdownMenuItem
-                        disabled={!userHasPermissionToUpdateFlow}
-                        onSelect={(e) => e.preventDefault()}
-                      >
-                        <div className="flex cursor-pointer  flex-row gap-2 items-center">
-                          <Trash2 className="h-4 w-4 text-destructive" />
-                          <span className="text-destructive">
-                            {t('Delete')}
-                          </span>
-                        </div>
-                      </DropdownMenuItem>
-                    </ConfirmationDeleteDialog>
-                  </PermissionNeededTooltip>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </div>
-          );
-        },
-      },
-    ],
-    [
-      doesUserHavePermissionToWriteFlow,
-      t,
-      selectedRows,
-      refresh,
-      isDropdownOpen,
-    ],
-  );
-
+  const bulkActions = useFlowsBulkActions({
+    selectedRows,
+    isDropdownOpen,
+    setIsDropdownOpen,
+    refresh,
+    setSelectedRows,
+    setRefresh,
+    refetch,
+  });
   return (
     <div className="flex flex-col gap-4 grow">
       <TaskLimitAlert />
